% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tempcyclesdata}
\alias{tempcyclesdata}
\title{Temperature cycling dataset from Wang & Dillon NCC 2014. doi:10.1038/nclimate2378}
\format{An object of class \code{data.frame} with 77181 rows and 27 columns.}
\source{
Wang & Dillon NCC 2014. doi:10.1038/nclimate2378
#' @examples
summary(tempcyclesdata)
if (require("dplyr")) {
tempcyclesdata %>%
  filter(period == "stdr") %>%
  group_by(region)
  summarise(mean_DTC = mean(DTC))
}
}
\usage{
tempcyclesdata
}
\description{
\pkg{tempcyclesdata} is dataset containing metadata, linear, and cycling data,
to be used with the \pkg{tempcycles} package. Only data passing all checks in included.
}
\keyword{datasets}

